py/agent/transformeragent/pyproject.toml
[project]
name = "transformeragent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.3.0",
    "datasets>=3.3.0",
    "evaluate>=0.4.3",
    "gradio-client>=1.7.0",
    "huggingface-hub>=0.28.1",
    "python-dotenv>=1.0.1",
    "torch>=2.6.0",
    "transformers[agents]>=4.48.3",
]
